:recycle: 도메인 서비스를 나눌 필요성이 없다고 판단되어 로직 제거
